Fix pointer-to-int-cast compiler warning
A void pointer(fast_memory) was type casted to uint64_t. This resulted
in a compiler warning [-Wpointer-to-int-cast]. This patch fixes the
warning.

+++ src/ethosu_driver.c
@@ -425,7 +425,7 @@ int ethosu_init(struct ethosu_driver *drv,
         }
     }
 
-    drv->fast_memory           = (uint64_t)fast_memory;
+    drv->fast_memory           = (uintptr_t)fast_memory;
     drv->fast_memory_size      = fast_memory_size;
     drv->power_request_counter = 0;
